Extract and restructure InvocationHandler for easier extension.
