+++ actors/test/actor-tests/src/test/java/cloud/orbit/actors/test/extension/InvocationHandlerTestExtension.java
@@ -35,12 +35,12 @@ import cloud.orbit.exception.UncheckedException;
 
 import java.lang.reflect.Method;
 import java.util.concurrent.atomic.AtomicBoolean;
-import java.util.concurrent.atomic.AtomicInteger;
+import java.util.concurrent.atomic.LongAdder;
 
 public class InvocationHandlerTestExtension implements InvocationHandlerExtension
 {
     private AtomicBoolean acceptCalls = new AtomicBoolean(true);
-    private AtomicInteger callCount = new AtomicInteger(0);
+    private LongAdder callCount = new LongAdder();
 
     @Override
     public Task beforeInvoke(final long startTimeNanos, final Object targetObject, final Method targetMethod, final Object[] params)
@@ -49,7 +49,7 @@ public class InvocationHandlerTestExtension implements InvocationHandlerExtensio
 
         if(!acceptCalls.get()) throw new UncheckedException("Not accepting calls");
 
-        callCount.incrementAndGet();
+        callCount.increment();
 
         return Task.done();
     }
@@ -61,7 +61,7 @@ public class InvocationHandlerTestExtension implements InvocationHandlerExtensio
 
         if(!acceptCalls.get()) throw new UncheckedException("Not accepting calls");
 
-        callCount.incrementAndGet();
+        callCount.increment();
 
         return Task.done();
     }
@@ -73,17 +73,17 @@ public class InvocationHandlerTestExtension implements InvocationHandlerExtensio
 
         if(!acceptCalls.get()) throw new UncheckedException("Not accepting calls");
 
-        callCount.incrementAndGet();
+        callCount.increment();
 
         return Task.done();
     }
 
     public void resetInvocationCount() {
-        callCount.set(0);
+        callCount.reset();
     }
 
-    public int getInvocationCount() {
-        return callCount.get();
+    public long getInvocationCount() {
+        return callCount.sum();
     }
 
     public void setAcceptCalls(final Boolean acceptCalls) {
